Update README.md (#134)
Hey, I really liked the project and was reading through the Readme.md file when I came across some spelling and grammatical errors that you might have missed while editing the documentation. It would be really a great opportunity for me if I could contribute to this project. Thank you.

Co-authored-by: Patrick von Platen <patrick.v.platen@gmail.com>